solver: split implementation from interface
