+++ control/control.go
@@ -9,6 +9,7 @@ import (
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/session/grpchijack"
 	"github.com/moby/buildkit/solver"
+	solverimpl "github.com/moby/buildkit/solver/solver"
 	"github.com/moby/buildkit/worker"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
@@ -25,13 +26,13 @@ type Opt struct {
 
 type Controller struct { // TODO: ControlService
 	opt    Opt
-	solver *solver.Solver
+	solver solver.Solver
 }
 
 func NewController(opt Opt) (*Controller, error) {
 	c := &Controller{
 		opt:    opt,
-		solver: solver.NewLLBSolver(opt.WorkerController, opt.Frontends),
+		solver: solverimpl.NewLLBSolver(opt.WorkerController, opt.Frontends),
 	}
 	return c, nil
 }
