+++ solver/solver/jobs.go
