+++ solver/solver/vertex.go
